requirements.txt
networkx==2.3
matplotlib==3.0.3
Cython==0.29.14
PyQt5==5.13.1
backports.functools_lru_cache==1.6.0
ipaddr==2.2.0
setuptools_svn==1.2
wincertstore==0.2
PyYAML==5.1.2
